requirements.txt
mkdocs>=1.5.0
mkdocs-material>=9.5.0
mkdocs-static-i18n>=1.2.0
PyYAML>=6.0
